requirements.txt
Django>=4.2,<4.3
wagtail>=5.0,<5.1
wagtailmenus
psycopg2-binary
django-environ
django_extensions
dj-database-url
django_cryptography
django_compressor
django-libsass
gunicorn
